Inter-nucleus distance calculations
For the diffusion equation, I need inter-nuclear distances.  Started on
the code for calculating that.  It may even work.  Had to reduce
position to 32 bit floats to get it to calculate properly (otherwise, we
run out of memory).  The piece-wise calculation of the distance matrix
is for a related Array Too Big error.

+++ PointClouds.py
@@ -108,7 +108,8 @@ class PointCloudReader(object):
                     pass
 
         if HAS_NUMPY:
-            posarray = np.zeros([len(all_data), 3, len(times)])
+            posarray = np.zeros([len(all_data), 3, len(times)],
+                                dtype=np.float32)
         else:
             posarray = [[[0 for k in times]
                          for j in range(3)]
